Move `read_geqdsk` into `geqdsk.py`
